requirements.txt
langchain
fastapi
uvicorn
jupyter
streamlit
PyPDF2
google-generativeai
faiss-cpu
langchain-google-genai
python-dotenv
langchain-community
langchain-google-genai
python-multipart